omr-evaluator/requirements.txt
huggingface_hub>=0.20.0
pillow>=10.0.0
requests>=2.31.0


